to: rename the Main comp to Home comp, chang some ui on it
